+++ src/allotropy/parsers/appbio_absolute_q/appbio_absolute_q_parser.py
@@ -47,7 +47,7 @@ from allotropy.parsers.vendor_parser import VendorParser
 class AppbioAbsoluteQParser(VendorParser):
     @property
     def display_name(self) -> str:
-        return "Applied Bio AbsoluteQ"
+        return "AppBio AbsoluteQ"
 
     @property
     def is_ready_to_use(self) -> bool: